requirements.txt
Flask==2.3.3
Flask-Limiter==3.5.0
python-dotenv==1.0.1
requests==2.31.0
MarkupSafe==2.1.5
gunicorn==21.2.0
flask-cors==4.0.0
